fix the CMakeLists file
This file is responsible to let the build system known which files we
want to be compiled in our project. So, notice that I've include the
files 'Article.cpp' 'PlanetArticle.cpp' into the executable. Also,
notice that we don't need to add the header files there, just the cpp
files.

As all the code is in the root directory, we don't need to call 'ADD_SUBDIRECTORY'.